+++ engines/wintermute/base/gfx/osystem/base_surface_osystem.h
@@ -89,7 +89,7 @@ private:
 	void genAlphaMask(Graphics::Surface *surface);
 	uint32 getPixelAt(Graphics::Surface *surface, int x, int y);
 
-	float _rotation;
+	uint32 _rotation;
 	bool _hasAlpha;
 	void *_lockPixels;
 	int _lockPitch;
